+++ modules/isd/include/FNormal.h
@@ -82,12 +82,6 @@ class IMPISDEXPORT FNormal : public OneDimensionalSufficientDistribution {
            exp(-square(FA_ - FM_) / (2 * square(sigma_)));
   }
 
-  IMPISD_DEPRECATED_FUNCTION_DECL(2.8)
-  virtual double density() const {
-    IMPISD_DEPRECATED_FUNCTION_DEF(2.8, "Use get_density() instead");
-    return get_density();
-  }
-
   /* change of parameters */
   void set_FA(double f) { FA_ = f; }
   void set_JA(double f) { JA_ = f; }